Migrate validation to Jekyll hooks (#2709)
Compared to the use of the validate.py scripts, this has the advantage of ensuring the validation is always performed.
It also gives access to enriched data, which is useful for validated generated content, such as links.

In addidition to checks already implemented in the previous 'validate.py' script, all links are checked (even those in the products description). This check takes time and is not performed by default, but it can be triggered by setting the MUST_CHECK_URLS environment variable to true. To ensure links are regularly verified, a new GitHub workflow has also been added. It runs every saturday at midnight.

Note that:

- some links had to be excluded because, despite being valid, they are sometime or systematically returning an error.
- a dependency to open-uri (https://github.com/ruby/open-uri) has been added to simplify the validation of HTTP links.
- the priority of the ProductDataEnricher plugin had to be set to normal, to allow hooks to be run before and after enrichment.

Some alternate URLs has been changed or removed due to an enforcement of the rules regarding permalinks and alternate_urls (no more dots).